requirements.txt
discord.py==1.6.0
requests==2.25.1
Pillow==8.1.2
